Revert "Refresh from upstream (#7)" (#8)
This reverts commit d7de26aed5885a6acb55b1a7aa0bdace3efa076f.

+++ test/framework/cluster_helpers.go
@@ -24,8 +24,10 @@ import (
 	. "github.com/onsi/gomega"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	clusterv1 "sigs.k8s.io/cluster-api/api/v1alpha4"
+	"k8s.io/apimachinery/pkg/runtime"
+	clusterv1 "sigs.k8s.io/cluster-api/api/v1alpha3"
 	"sigs.k8s.io/cluster-api/test/framework/internal/log"
+	"sigs.k8s.io/cluster-api/test/framework/options"
 	"sigs.k8s.io/cluster-api/util/patch"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -34,7 +36,7 @@ import (
 type CreateClusterInput struct {
 	Creator      Creator
 	Cluster      *clusterv1.Cluster
-	InfraCluster client.Object
+	InfraCluster runtime.Object
 }
 
 // CreateCluster will create the Cluster and InfraCluster objects.
@@ -61,7 +63,7 @@ type GetAllClustersByNamespaceInput struct {
 	Namespace string
 }
 
-// GetAllClustersByNamespace returns the list of Cluster object in a namespace.
+// GetAllClustersByNamespace returns the list of Cluster object in a namespace
 func GetAllClustersByNamespace(ctx context.Context, input GetAllClustersByNamespaceInput) []*clusterv1.Cluster {
 	clusterList := &clusterv1.ClusterList{}
 	Expect(input.Lister.List(ctx, clusterList, client.InNamespace(input.Namespace))).To(Succeed(), "Failed to list clusters in namespace %s", input.Namespace)
@@ -80,7 +82,7 @@ type GetClusterByNameInput struct {
 	Namespace string
 }
 
-// GetClusterByName returns a Cluster object given his name.
+// GetClusterByName returns a Cluster object given his name
 func GetClusterByName(ctx context.Context, input GetClusterByNameInput) *clusterv1.Cluster {
 	cluster := &clusterv1.Cluster{}
 	key := client.ObjectKey{
@@ -142,6 +144,9 @@ type DeleteClusterInput struct {
 
 // DeleteCluster deletes the cluster and waits for everything the cluster owned to actually be gone.
 func DeleteCluster(ctx context.Context, input DeleteClusterInput) {
+	if options.SkipResourceCleanup {
+		return
+	}
 	By(fmt.Sprintf("Deleting cluster %s", input.Cluster.GetName()))
 	Expect(input.Deleter.Delete(ctx, input.Cluster)).To(Succeed())
 }
@@ -154,6 +159,9 @@ type WaitForClusterDeletedInput struct {
 
 // WaitForClusterDeleted waits until the cluster object has been deleted.
 func WaitForClusterDeleted(ctx context.Context, input WaitForClusterDeletedInput, intervals ...interface{}) {
+	if options.SkipResourceCleanup {
+		return
+	}
 	By(fmt.Sprintf("Waiting for cluster %s to be deleted", input.Cluster.GetName()))
 	Eventually(func() bool {
 		cluster := &clusterv1.Cluster{}
